+++ src/one_app_to_rule_them_all.py
@@ -47,7 +47,7 @@ class Task:
 
         self.task_color = COLOR_PALETTE[Task.task_number]
         Task.task_number += 1
-        self.task_name = f"Z:{Task.task_number}"
+        self.task_name = f"Z{Task.task_number}"
 
     def __str__(self):
         return self.task_name